Added working converter from old datasets (not tested on signals yet).. Moved profile types into the profile collection, so only one segment map is maintained. Removed border rectangles as fields in the CellularComponent, and changed position array to int[]. Old datasets can still be opened directly if needed.

+++ src/analysis/nucleus/FileProcessingTask.java
@@ -121,7 +121,7 @@ public class FileProcessingTask  extends AbstractProgressAction  {
 		  Nucleus n 			 = cell.getNucleus();
 		  PolygonRoi nucleus 	 = new PolygonRoi(n.createPolygon(), PolygonRoi.POLYGON);
 		  
-		  double[] position = n.getPosition();
+		  int[] position = n.getPosition();
 		  nucleus.setLocation(position[CellularComponent.X_BASE],position[CellularComponent.Y_BASE]); // translate the roi to the image coordinates
 	}
