+++ src/components/nuclei/sperm/PigSpermNucleus.java
@@ -68,11 +68,11 @@ public class PigSpermNucleus
     	super();
     }
 
-    public PigSpermNucleus (Roi roi, File file, int number, double[] position) { // construct from an roi
+    public PigSpermNucleus (Roi roi, File file, int number, int[] position) { // construct from an roi
 		super(roi, file, number, position);
 	}
     
-    public PigSpermNucleus (Roi roi, File file, int number, double[] position, IPoint centreOfMass) { // construct from an roi
+    public PigSpermNucleus (Roi roi, File file, int number, int[] position, IPoint centreOfMass) { // construct from an roi
 		super(roi, file, number, position, centreOfMass);
 	}
